1.2

+++ structs.go
@@ -17,11 +17,11 @@ type Credits struct {
 }
 
 type Statistics struct {
-	Average_successful_investment   float32 `json:"average_successful_investment"`
-	Average_unsuccessful_investment float32 `json:"average_unsuccessful_investment"`
-	Total_assignments_made          int64   `json:"total_assignments_made"`
-	Total_successful_assignments    int64   `json:"total_successful_assignments"`
-	Total_unsuccessful_assignments  int64   `json:"total_unsuccessful_assignments"`
+	Average_successful_investment   int64 `json:"average_successful_investment"`
+	Average_unsuccessful_investment int64 `json:"average_unsuccessful_investment"`
+	Total_assignments_made          int64 `json:"total_assignments_made"`
+	Total_successful_assignments    int64 `json:"total_successful_assignments"`
+	Total_unsuccessful_assignments  int64 `json:"total_unsuccessful_assignments"`
 }
 
 type Average struct {
